+++ src/api/index.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { JoinData, LoginData, Email, ReservationPostData } from './type';
+import { ReservationPostData } from './type';
 import { getCookie } from '../utils/utils';
 import { AddReviewData } from '../@types/interface';
 
@@ -33,37 +33,6 @@ clientToken.interceptors.request.use(
   },
   error => Promise.reject(error),
 );
-
-export const postLogin = async (loginData: LoginData) => {
-  const res = await client.post('/login', loginData);
-  return res.data;
-};
-
-export const postJoin = async (joinData: JoinData) => {
-  const res = await client.post('/user/signup', joinData);
-  return res.data;
-};
-
-export const postEmail = async (email: Email) => {
-  const res = await client.post('/user/email/confirmation', email);
-  return res.data;
-};
-
-export const getVerify = async (verify: string, email: string) => {
-  const res = await client.get(`/user/verify/${verify}?email=${email}`);
-  return res.data;
-};
-
-// 토큰 쓰는 API 형식
-export const postLogout = async () => {
-  const res = await clientToken.post('/logout');
-  return res.data;
-};
-
-export const testToken = async () => {
-  const res = await clientToken.get('/user/test');
-  return res;
-};
 
 // 장바구니 가져오기
 export const getBasket = async () => {
